pyproject.toml
[tool.poetry]
name = "shorturl_backend"
version = "0.1.0"
description = ""
authors = ["alan <alancmp0304@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.8"
Flask = "^2.0.2"
python-dotenv = "^0.19.2"
pymongo = "^4.0"
redis = "^4.0.2"
requests = "^2.26.0"
PyJWT = "^2.3.0"
passlib = "^1.7.4"
pydantic = {extras = ["email"], version = "^1.8.2"}
pytz = "^2021.3"
bcrypt = "^3.2.0"
gunicorn = {extras = ["gevent"], version = "^20.1.0"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
